+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Current development]
 * Added new functions date_tz and strtotime_tz
 * Rework GetAmazonCustomerProfileElement to read permissions from Amazon Platform Config
+* Added Alexa Skill Certification status check
 
 
 ## [Releases]